Refactor SMFLogger integration
- Improve code structure
- reinroduce "self.isInitialized" to enable warnings about a not setup state
- Hide shared instance to other classes
- Remove debug config support as the usage leads to wrong expectations (build release to test)
